mostly finished everything except drag
still bugs with edit.
creating cards on boards other than done and then moving those cards to the done board with the done button and then moving one of those cards to a diff board with the edit button causes a bug, where when i try to edit a card on the done board causes me to edit the aforementioned card's board i moved.
i.e.
there are 3 cards on the done board
i move one to the stuck board
i try to edit the 1st card on the done board - i end up editing the 1st card on the stuck board
i try to edit 2/3 card on the done board - i get an error since i try to edit a non existent 2/3 card on the stuck board